Moved makeQueenTable to table-component
