+++ src/queen-detail/table-component.js
@@ -1,0 +1,17 @@
+export function makeQueenTable(queen) {
+    const html = /*html*/`
+        <dl>
+            <dt>Name:</dt>
+            <dd>${queen.name}</dd>
+            <dt>Winner:</dt>
+            <dd>${queen.winner ? 'Yes' : 'No'}</dd>
+            <dt>Miss Congeniality:</dt>
+            <dd>${queen.missCongeniality ? 'Yes' : 'No'}</dd>
+            <dt>Quote:</dt>
+            <dd>${queen.quote}</dd>
+        </dl>
+    `;
+    const template = document.createElement('template');
+    template.innerHTML = html;
+    return template.content;
+}
