+++ test/table-component.test.js
@@ -1,24 +1,8 @@
+import { makeQueenTable } from '../src/queen-detail/table-component.js';
+
 const test = QUnit.test;
 
 QUnit.module('Table component');
-
-function makeQueenTable(queen) {
-    const html = /*html*/`
-        <dl>
-            <dt>Name:</dt>
-            <dd>${queen.name}</dd>
-            <dt>Winner:</dt>
-            <dd>${queen.winner ? 'Yes' : 'No'}</dd>
-            <dt>Miss Congeniality:</dt>
-            <dd>${queen.missCongeniality ? 'Yes' : 'No'}</dd>
-            <dt>Quote:</dt>
-            <dd>${queen.quote}</dd>
-        </dl>
-    `;
-    const template = document.createElement('template');
-    template.innerHTML = html;
-    return template.content;
-}
 
 test('make table', function(assert) {
     // arrange
